Move tests to client
