+++ testkit/build.sbt
@@ -6,7 +6,6 @@ libraryDependencies ++= Seq(
   //  "com.typesafe.akka" %% "akka-actor" % akkaVersion
   "com.typesafe" % "config" % "1.3.0",
 
-  "org.apache.kafka" % "kafka-clients" % versions.kafka,
   "org.slf4j" % "slf4j-api" % versions.slf4j,
   "org.slf4j" % "log4j-over-slf4j" % versions.slf4j,
   "ch.qos.logback" % "logback-classic" % "1.1.3",
@@ -15,6 +14,6 @@ libraryDependencies ++= Seq(
       exclude("org.slf4j", "slf4j-log4j12"),
 
   //Test deps
-  "org.scalatest" % "scalatest_2.11" % versions.scalaTest % "test",
+//  "org.scalatest" % "scalatest_2.11" % versions.scalaTest % "test",
   "org.apache.curator" % "curator-test" % "2.7.0" //3.0.0
 )
